Refactor archivos server game logic
