+++ server/CMakeLists.txt
@@ -2,27 +2,27 @@ target_sources(taller_server
     PRIVATE
     # .cpp files
         main.cpp
-        server.cpp
-        client_handler.cpp
-        player.cpp
-        map.cpp
-        weapon/weapon_knife.cpp
-        weapon/fire_weapon.cpp
-        weapon/weapon_ak47.cpp
-        match.cpp
+        src/server.cpp
+        src/client_handler.cpp
+        src/player.cpp
+        src/map.cpp
+        src/weapon/weapon_knife.cpp
+        src/weapon/fire_weapon.cpp
+        src/weapon/weapon_ak47.cpp
+        src/match.cpp
     PUBLIC
     # .h files
-        server.h
-        client_handler.h
-        player.h
-        map.h
-        weapon/weapon_knife.h
-        weapon/weapon.h
-        weapon/fire_weapon.h
-        weapon/weapon_ak47.h
-        match.h
-        enums/player_type.h
-        enums/player_state.h
-        enums/weapon_type.h
-        enums/match_phase.h
+        include/server.h
+        include/client_handler.h
+        include/player.h
+        include/map.h
+        include/weapon/weapon_knife.h
+        include/weapon/weapon.h
+        include/weapon/fire_weapon.h
+        include/weapon/weapon_ak47.h
+        include/match.h
+        include/enums/player_type.h
+        include/enums/player_state.h
+        include/enums/weapon_type.h
+        include/enums/match_phase.h
     )
