+++ server/include/enums/player_type.h
