+++ server/include/enums/weapon_type.h
