+++ server/include/weapon/weapon_ak47.h
