--- server/server.h
@@ -1,25 +1,0 @@
-#ifndef SERVER
-#define SERVER
-
-#include <string>
-#include <vector>
-
-#include "common/socket.h"
-#include "client_handler.h"
-
-class Server {
-private:
-    std::string port;
-    Socket srv;
-    std::vector<ClientHandler*> handlers;
-
-public:
-    explicit Server(const std::string& port);
-
-    void run();
-
-    ~Server() = default;
-
-};
-
-#endif
