+++ server/src/weapon/weapon_ak47.cpp
@@ -1,3 +1,3 @@
-#include "weapon_ak47.h"
+#include "../../include/weapon/weapon_ak47.h"
 
 WeaponAk47::WeaponAk47() : FireWeapon(35, 2700, 30, 600){}
