+++ server/src/weapon/weapon_awp.cpp
@@ -1,3 +1,3 @@
-#include "weapon_awp.h"
+#include "../../include/weapon/weapon_awp.h"
 
 WeaponAwp::WeaponAwp() : FireWeapon(50, 3500, 10, 1000){}
