+++ server/src/weapon/weapon_glock.cpp
@@ -1,3 +1,3 @@
-#include "weapon_glock.h"
+#include "../../include/weapon/weapon_glock.h"
 
 WeaponGlock::WeaponGlock() : FireWeapon(45, 0, 40, 700){}
