+++ server/src/weapon/weapon_knife.cpp
@@ -1,0 +1,5 @@
+#include "../../include/weapon/weapon_knife.h"
+
+int WeaponKnife::use() {
+    return 20;
+}
