+++ server/src/weapon/weapon_m3.cpp
@@ -1,3 +1,3 @@
-#include "weapon_m3.h"
+#include "../../include/weapon/weapon_m3.h"
 
 WeaponM3::WeaponM3() : FireWeapon(70, 3000, 20, 500){}
