--- server/weapon/weapon_knife.cpp
@@ -1,5 +1,0 @@
-#include "weapon_knife.h"
-
-int WeaponKnife::use() {
-    return 20;
-}
